gedbrowser-renderer/src/main/java/org/schoellerfamily/gedbrowser/renderer/SubmitterLinkPhraseRenderer.java
package org.schoellerfamily.gedbrowser.renderer;

/**
 * @author Dick Schoeller
 */
public class SubmitterLinkPhraseRenderer implements PhraseRenderer {
    /**
     * Holder for the SubmitterLinkRenderer that is using this helper.
     */
    private final transient SubmitterLinkRenderer slRenderer;

    /**
     * Constructor.
     *
     * @param renderer the renderer that this is associated with.
     */
    protected SubmitterLinkPhraseRenderer(
            final SubmitterLinkRenderer renderer) {
        this.slRenderer = renderer;
    }

    /**
     * {@inheritDoc}
     */
    @Override
    public final String renderAsPhrase() {
        return slRenderer.getNameString();
    }
}
